separate filterreports function from slackreports to filter.js and renamed into findreportstobesent
