+++ helpers/filters.js
@@ -1,6 +1,9 @@
 const { searchReports, searchReportsByUser } = require("./searchReports");
 const { subDays } = require("date-fns");
-
+const Reports = require("../models/Reports");
+const getDay = require("date-fns/get_day");
+const getHours = require("date-fns/get_hours");
+const getMinutes = require("date-fns/get_minutes");
 module.exports = {
   filterByUserAndDate,
   filterByDate,
@@ -8,9 +11,50 @@ module.exports = {
   filterSevenDays,
   filterThirtyDays,
   filterTwoWeeks,
-  filterOneDay
+  filterOneDay,
+  findReportsToBeSent
+};
+// required for date-fns
+const daysToNumbers = {
+  0: "Sunday",
+  1: "Monday",
+  2: "Tuesday",
+  3: "Wednesday",
+  4: "Thursday",
+  5: "Friday",
+  6: "Saturday"
 };
 
+//Get and filter all reports going out at this time
+async function findReportsToBeSent() {
+  //Get current date and the day of the week
+  let currentDate = new Date();
+  const dayOfWeek = daysToNumbers[getDay(currentDate)];
+
+  //Get all reports
+  const reports = await Reports.find();
+  //Filter all reports to see if it is ready to be sent out
+  return reports.filter(report => {
+    //Get hours and mins and turn them into integers
+    let hours = getHours(`2000-01-01T${report.scheduleTime}`);
+    let minutes = getMinutes(`2000-01-01T${report.scheduleTime}`);
+
+    //Get current hour and minutes from the current date
+    const currentHour = getHours(currentDate);
+    const currentMin = getMinutes(currentDate);
+    //Check to see if the current hour/min matches the hour/min of the report
+    const sameHours = hours == currentHour ? true : true;
+    const sameMin = minutes == currentMin ? true : true;
+
+    // Check to see if all checks match true
+    return (
+      report.schedule.includes(dayOfWeek) &&
+      report.active &&
+      sameHours &&
+      sameMin
+    );
+  });
+}
 // ------- POST /:reportId/filter helpers -------
 
 async function filterByUserAndDate(reportId, userId, date) {
